Refactor axis svg/view
